split main package into two files (later there will be a server.go file, too)
